+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [2.3.1] - 2023-XX-XX
+- Fix issue [#77](https://github.com/intersystems/language-server/issues/77): Incorrect 'Class/Routine/Include file does not exist' Diagnostics after namespace switch
 - Fix issue [#261](https://github.com/intersystems/language-server/issues/261): Add parser support for CSS @keyframes rule
 
 ## [2.3.0] - 2023-02-13